requirements.txt
selenium
webdriver-manager
requests
psycopg2-binary
pandas
streamlit
schedule